Add verbose for test debugging
